+++ src/scorer/toot/followed_accounts_scorer.ts
@@ -1,0 +1,27 @@
+/*
+ * One point if you follow the account (retoots by followed accounts are picked up by the
+ * RetootsInFeedScorer).
+ */
+import Account from '../../api/objects/account';
+import MastoApi from '../../api/api';
+import Toot from '../../api/objects/toot';
+import TootScorer from '../toot_scorer';
+import { ScoreName } from '../../enums';
+import { type StringNumberDict } from '../../types';
+
+
+export default class FollowedAccountsScorer extends TootScorer {
+    description = "Favour accounts you follow";
+
+    constructor() {
+        super(ScoreName.FOLLOWED_ACCOUNTS);
+    };
+
+    async prepareScoreData(): Promise<StringNumberDict> {
+        return Account.countAccounts(await MastoApi.instance.getFollowedAccounts());
+    };
+
+    async _score(toot: Toot): Promise<number> {
+        return this.scoreData[toot.account.webfingerURI] ?? 0;
+    }
+};