+++ src/scorer/toot/num_retoots_scorer.ts
@@ -1,0 +1,19 @@
+/*
+ * Score how many times the toot has been retooted.
+ */
+import Toot from '../../api/objects/toot';
+import TootScorer from '../toot_scorer';
+import { ScoreName } from '../../enums';
+
+
+export default class NumRetootsScorer extends TootScorer {
+    description = "Favour toots that are retooted a lot";
+
+    constructor() {
+        super(ScoreName.NUM_RETOOTS);
+    }
+
+    async _score(toot: Toot) {
+        return toot.realToot.reblogsCount || 0;
+    }
+};