+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # NEXT RELEASE
 
+### v0.72.8
+* Rename `FeatureScorer` to `TootScorer`
+* Rename `HOMESERVER_TIMELINE_TOOTS` to just `HOMESERVER_TOOTS`
+* Remove moreOldToots argument from triggerFeedUpate() (client app should call triggerHomeTimelineBackFill())
+* Remove config `hashtagTootRetrievalDelaySeconds`; don't sleep before launching hashtag `Toot` retrieval
+* Use `fetchGenerator() for all API calls instead of either `fetch()` or `fetchGenerator()`
+* Lower config.api.backgroundLoadSleepBetweenRequestsMS to 1000
+
 ### v0.72.7
-* Mark Toot objects that are missing a trendingLinks property as needing to be recompleted
+* Mark `Toot` objects that are missing a `trendingLinks` property as needing to be recompleted
 
 ### v0.72.6
 * Use appropriate conditional types in `MastoApi`, rename `MastodonApiObj` to `ApiObj`